requirements.txt
# Core packages
pandas
python-dotenv
requests

# Email
smtplib
email

# Google Sheets access
gspread
oauth2client

# OpenAI API
openai>=1.0.0

# Streamlit UI
streamlit

# Optional (used in watch script for desktop notifications on macOS)
osascript

# File watching (optional)
watchdog
